remove main html tag in Blog
From MDN[1] : The content of a <main> element should be unique to the
document. Content that is repeated across a set of documents or document
sections such as sidebars, navigation links, copyright information, site
logos, and search forms shouldn't be included unless the search form is
the main function of the page.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/main#usage_notes